DEP-61 refactor : 토큰 유효성 검증과, payload 디코딩 책임을 클래스로 분리
